moved common method from soluble phosphorus to layer data
